Organised Naives_Bayes and fixed small errors
